fix summary and make ReporterForTest use verbose flag
