+++ dartdoc_test/example/lib/example.dart
@@ -16,8 +16,9 @@ import 'dart:convert';
 
 /// Example of documentation comments and code samples in Dart.
 
-/// code block without specified language
-/// ```
+/// Dart code block will be analyzed by dartdoc_test.
+///
+/// ```dart
 /// final result = add(2, 3);
 /// print(result); // 5
 /// ```
@@ -25,7 +26,18 @@ int add(int a, int b) {
   return a + b;
 }
 
-/// multiple code blocks
+/// If you don't specify the language for code blocks, it still be analyzed.
+///
+/// ```
+/// final result = subtract(5, 3);
+/// print(result); // 2
+/// ```
+int subtract(int x, int y) {
+  return x - y;
+}
+
+/// Multiple code blocks are also analyzed.
+///
 /// ```dart
 /// final result = multiply(4, 5);
 /// print(result); // 20
@@ -39,7 +51,7 @@ int multiply(int x, int y) {
   return x * y;
 }
 
-/// non-dart code block will be ignored.
+/// Non-Dart code blocks are ignored.
 ///
 /// ```python
 /// def add(a, b):
@@ -52,7 +64,7 @@ int multiply(int x, int y) {
 /// ```
 void ignore() {}
 
-/// you can ignore code sample by using `no-test` tag.
+/// You can ignore code sample by using `dartdoc_test:ignore_error` tag.
 ///
 /// ```dart
 /// // dartdoc_test:ignore_error
@@ -60,54 +72,3 @@ void ignore() {}
 /// tagIgnore() // it is not reported.
 /// ```
 void tagIgnore() {}
-
-/// Should return: Expected to find ';'.
-/// ```dart
-/// final fact = factorial(5)
-/// print(fact); // 120
-/// ```
-int factorial(int n) {
-  if (n < 0) throw ArgumentError('Negative numbers are not allowed.');
-  return n == 0 ? 1 : n * factorial(n - 1);
-}
-
-/// Checks if a string is a palindrome.
-///
-/// This method ignores case and non-alphanumeric characters.
-///
-/// Example:
-/// ```dart
-/// final isPalindrome = isPalindrome('A man, a plan, a canal, Panama');
-/// print(isPalindrome); // true
-/// ```
-bool isPalindrome(String s) {
-  var sanitized = s.replaceAll(RegExp(r'[^A-Za-z0-9]'), '').toLowerCase();
-  return sanitized == sanitized.split('').reversed.join('');
-}
-
-/// Should return: Local variable 'gcd' can't be referenced before it is declared.
-/// ```dart
-/// final gcd = gcd(48, 18);
-/// print(gcd); // 6
-/// ```
-int gcd(int a, int b) {
-  while (b != 0) {
-    var t = b;
-    b = a % b;
-    a = t;
-  }
-  return a;
-}
-
-/// Convert a string to a list of lines.
-///
-/// This method splits a string into lines using the newline character.
-///
-/// Example:
-/// ```dart
-/// final lines = splitLines('Hello\nWorld');
-/// print(lines); // ['Hello', 'World']
-/// ```
-List<String> splitLines(String s) {
-  return LineSplitter.split(s).toList();
-}
