--- dartdoc_test/example/lib/main.dart
@@ -1,33 +1,0 @@
-// Copyright 2023 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/// Simple function that prints `"Hello world!"`.
-///
-/// **Example**
-/// ```dart
-/// sayHello(); // prints: Hello world!
-/// ```
-void sayHello() {
-  // My comment
-  print('Hello world!');
-}
-
-/// **Example**
-/// ```dart
-/// foo(); // prints: foo!
-/// ```
-void foo() {
-  // My comment
-  print('foo!');
-}
