+++ dartdoc_test/lib/src/logger.dart
@@ -62,7 +62,7 @@ class Summary {
   factory Summary.from(List<DartdocAnalysisResult> results) {
     final isFailed = results.indexWhere((r) => r.errors.isNotEmpty) != -1;
     final errors =
-        results.expand((r) => r.errors).where((e) => e.generatedSpan != null);
+        results.expand((r) => r.errors).where((e) => e.commentSpan != null);
     final samples = results.map((r) => r.file);
     final files =
         samples.map((s) => s.sample.comment.span.sourceUrl?.path).toSet();
