+++ dartdoc_test/test/testdata/codeSample/error_example_0.dart
@@ -1,0 +1,7 @@
+
+import '../../../example/lib/error_example.dart';
+
+void main() {
+  final fact = factorial(5)
+  print(fact); // 120
+}
