+++ dartdoc_test/test/testdata/codeSample/error_example_2.dart
@@ -1,0 +1,7 @@
+
+import '../../../example/lib/error_example.dart';
+
+void main() {
+  final gcd = gcd(48, 18);
+  print(gcd); // 6
+}
