+++ dartdoc_test/test/testdata/codeSample/error_example_3.dart
@@ -1,0 +1,7 @@
+
+import '../../../example/lib/error_example.dart';
+
+void main() {
+  final length = 5.length;
+  print(length); // 1
+}
