+++ dartdoc_test/test/testdata/codeSample/example_2.dart
@@ -2,6 +2,6 @@ import 'dart:convert';
 import '../../../example/lib/example.dart';
 
 void main() {
-  final result = multiply(6, 7);
-  print(result); // 42
+  final result = multiply(4, 5);
+  print(result); // 20
 }
