+++ dartdoc_test/test/testdata/codeSample/example_3.dart
@@ -2,6 +2,6 @@ import 'dart:convert';
 import '../../../example/lib/example.dart';
 
 void main() {
-  final fact = factorial(5)
-  print(fact); // 120
+  final result = multiply(6, 7);
+  print(result); // 42
 }
