--- dartdoc_test/test/testdata/codeSample/example_6.dart
@@ -1,7 +1,0 @@
-import 'dart:convert';
-import '../../../example/lib/example.dart';
-
-void main() {
-  final gcd = gcd(48, 18);
-  print(gcd); // 6
-}
