--- dartdoc_test/test/testdata/codeSample/example_7.dart
@@ -1,7 +1,0 @@
-import 'dart:convert';
-import '../../../example/lib/example.dart';
-
-void main() {
-  final lines = splitLines('Hello\nWorld');
-  print(lines); // ['Hello', 'World']
-}
