--- dartdoc_test/test/testdata/codeSample/main_0.dart
@@ -1,6 +1,0 @@
-
-import '../../../example/lib/main.dart';
-
-void main() {
-  sayHello(); // prints: Hello world!
-}
