--- dartdoc_test/test/testdata/codeSample/main_1.dart
@@ -1,6 +1,0 @@
-
-import '../../../example/lib/main.dart';
-
-void main() {
-  foo(); // prints: foo!
-}
