+++ dartdoc_test/test/testdata/run analyze command by default with verbose flag.txt
@@ -1,36 +1,41 @@
 $ dart run bin/dartdoc_test.dart -v
 Extracting code samples ...
 Analyzing code samples ...
-.dartdoc_test/example_0.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-.dartdoc_test/example_1.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-.dartdoc_test/example_2.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-original error: .dartdoc_test/example_4.dart:5:27: Expected to find ';'.
-example/lib/example.dart:66:29: : Expected to find ';'.
+original error: .dartdoc_test/error_example_0.dart:5:27: Expected to find ';'.
+example/lib/error_example.dart:17:29: : Expected to find ';'.
    ╷
-66 │ /// final fact = factorial(5)
+17 │ /// final fact = factorial(5)
    │                             ^
    ╵
 
-.dartdoc_test/example_4.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-original error: .dartdoc_test/example_5.dart:5:24: Local variable 'isPalindrome' can't be referenced before it is declared.
-example/lib/example.dart:80:11: : Local variable 'isPalindrome' can't be referenced before it is declared.
+original error: .dartdoc_test/error_example_1.dart:5:24: Local variable 'isPalindrome' can't be referenced before it is declared.
+example/lib/error_example.dart:31:11: : Local variable 'isPalindrome' can't be referenced before it is declared.
    ╷
-80 │ /// final isPalindrome = isPalindrome('A man, a plan, a canal, Panama');
+31 │ /// final isPalindrome = isPalindrome('A man, a plan, a canal, Panama');
    │           ^^^^^^^^^^^^
    ╵
 
-.dartdoc_test/example_5.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-.dartdoc_test/example_5.dart:2:8: Unused import: '../example/lib/example.dart'. (ignored because issue occurs in the generated code)
-original error: .dartdoc_test/example_6.dart:5:15: Local variable 'gcd' can't be referenced before it is declared.
-example/lib/example.dart:88:36: : Local variable 'gcd' can't be referenced before it is declared.
+.dartdoc_test/error_example_1.dart:2:8: Unused import: '../example/lib/error_example.dart'. (ignored because issue occurs in the generated code)
+original error: .dartdoc_test/error_example_2.dart:5:15: Local variable 'gcd' can't be referenced before it is declared.
+example/lib/error_example.dart:39:36: : Local variable 'gcd' can't be referenced before it is declared.
    ╷
-88 │ /// Should return: Local variable 'gcd' can't be referenced before it is declared.
+39 │ /// Should return: Local variable 'gcd' can't be referenced before it is declared.
    │                                    ^^^
    ╵
 
-.dartdoc_test/example_6.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
-.dartdoc_test/example_6.dart:2:8: Unused import: '../example/lib/example.dart'. (ignored because issue occurs in the generated code)
-.dartdoc_test/example_7.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
+.dartdoc_test/error_example_2.dart:2:8: Unused import: '../example/lib/error_example.dart'. (ignored because issue occurs in the generated code)
+original error: .dartdoc_test/error_example_3.dart:5:20: The getter 'length' isn't defined for the type 'int'.
+example/lib/error_example.dart:53:32: : The getter 'length' isn't defined for the type 'int'.
+   ╷
+53 │ /// Should return: The getter 'length' isn't defined for the class 'int'.
+   │                                ^^^^^^
+   ╵
+
+.dartdoc_test/error_example_3.dart:2:8: Unused import: '../example/lib/error_example.dart'. (ignored because issue occurs in the generated code)
+.dartdoc_test/example_0.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
+.dartdoc_test/example_1.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
+.dartdoc_test/example_2.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
+.dartdoc_test/example_3.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
 .dartdoc_test/extractor_0.dart:1:8: Unused import: 'dart:convert'. (ignored because issue occurs in the generated code)
 .dartdoc_test/extractor_0.dart:2:8: Unused import: 'dart:io'. (ignored because issue occurs in the generated code)
 .dartdoc_test/extractor_0.dart:3:8: Unused import: 'package:path/path.dart'. (ignored because issue occurs in the generated code)
@@ -55,4 +60,4 @@ example/lib/example.dart:88:36: : Local variable 'gcd' can't be referenced befor
 .dartdoc_test/extractor_2.dart:6:8: Unused import: 'package:analyzer/dart/ast/visitor.dart'. (ignored because issue occurs in the generated code)
 .dartdoc_test/extractor_2.dart:7:8: Unused import: 'package:markdown/markdown.dart'. (ignored because issue occurs in the generated code)
 .dartdoc_test/extractor_2.dart:8:8: Unused import: 'package:source_span/source_span.dart'. (ignored because issue occurs in the generated code)
-FAILED: 36 issues found (Found 12 code samples in 3 files)
+FAILED: 4 issues found (Found 11 code samples in 3 files)
