+++ dartdoc_test/test/testdata/run analyze command.txt
@@ -1,22 +1,28 @@
 $ dart run bin/dartdoc_test.dart analyze
 Extracting code samples ...
 Analyzing code samples ...
-example/lib/example.dart:66:29: : Expected to find ';'.
+example/lib/error_example.dart:17:29: : Expected to find ';'.
    ╷
-66 │ /// final fact = factorial(5)
+17 │ /// final fact = factorial(5)
    │                             ^
    ╵
 
-example/lib/example.dart:80:11: : Local variable 'isPalindrome' can't be referenced before it is declared.
+example/lib/error_example.dart:31:11: : Local variable 'isPalindrome' can't be referenced before it is declared.
    ╷
-80 │ /// final isPalindrome = isPalindrome('A man, a plan, a canal, Panama');
+31 │ /// final isPalindrome = isPalindrome('A man, a plan, a canal, Panama');
    │           ^^^^^^^^^^^^
    ╵
 
-example/lib/example.dart:88:36: : Local variable 'gcd' can't be referenced before it is declared.
+example/lib/error_example.dart:39:36: : Local variable 'gcd' can't be referenced before it is declared.
    ╷
-88 │ /// Should return: Local variable 'gcd' can't be referenced before it is declared.
+39 │ /// Should return: Local variable 'gcd' can't be referenced before it is declared.
    │                                    ^^^
    ╵
 
-FAILED: 36 issues found (Found 12 code samples in 3 files)
+example/lib/error_example.dart:53:32: : The getter 'length' isn't defined for the type 'int'.
+   ╷
+53 │ /// Should return: The getter 'length' isn't defined for the class 'int'.
+   │                                ^^^^^^
+   ╵
+
+FAILED: 4 issues found (Found 11 code samples in 3 files)
